+++ src/main/kotlin/org/wisp/gatesawakened/Di.kt
@@ -3,10 +3,12 @@ package org.wisp.gatesawakened
 import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.SettingsAPI
 import com.fs.starfarer.api.campaign.SectorAPI
+import com.fs.starfarer.api.campaign.comm.IntelManagerAPI
 import org.wisp.gatesawakened.logging.DebugLogger
 
 class Di(
     val sector: SectorAPI = Global.getSector(),
+    val intelManager: IntelManagerAPI = sector.intelManager,
     val settings: SettingsAPI = Global.getSettings(),
     val logger: DebugLogger = Global.getLogger(Di::class.java)
 )