+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.1
+
+- Added new intel that displays all active gates
+- Updated intel icons to be cooler
+- Reduced fuel requirement to use gates to 80%
+- Removed dependency on LazyLib
+
+- Fixed bug where the second quest would not appear, even if you met the requirements
+
 ## 1.0.1 (2019-9-12)
 
 - Fixed jump using double the stated fuel price.